+++ web/app/themes/rebe/functions.php
@@ -24,6 +24,32 @@ add_action('after_switch_theme', function () {
 });
 
 /**
- * Require composer autoloader
+ * Sage includes
+ *
+ * The $sage_includes array determines the code library included in your theme.
+ * Add or remove files to the array as needed. Supports child theme overrides.
+ *
+ * Please note that missing files will produce a fatal error.
+ *
+ * @link https://github.com/roots/sage/pull/1042
  */
-require_once __DIR__ . '/vendor/autoload.php';
+$sage_includes = [
+    'src/helpers.php',    // Helper functions
+    'src/setup.php',      // Theme setup
+    'src/filters.php',    // Filters
+    'src/admin.php',      // Admin
+    'src/api.php'         // API
+];
+foreach ($sage_includes as $file) {
+    if (!$filepath = locate_template($file)) {
+        trigger_error(sprintf(__('Error locating %s for inclusion', 'rebe'), $file), E_USER_ERROR);
+    }
+    require_once $filepath;
+}
+unset($file, $filepath);
+/**
+ * Require Composer autoloader if installed on it's own
+ */
+if (file_exists($composer = __DIR__ . '/vendor/autoload.php')) {
+    require_once $composer;
+}